requirements.txt
# Core machine learning libraries
torch>=2.0.0,<3.0.0
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0

# Hugging Face ecosystem
huggingface_hub>=0.16.0,<1.0.0

# Deep learning utilities
einops>=0.6.0,<1.0.0
tqdm>=4.64.0,<5.0.0

# Data sources
python-binance>=1.0.0,<2.0.0
akshare>=1.9.0,<2.0.0

# Templating for static site generation
jinja2>=3.0.0,<4.0.0

# Visualization
matplotlib>=3.5.0,<4.0.0

# Task scheduling
schedule>=1.2.0,<2.0.0